Implement reminder history
Reminders can now be archived. Archived reminders will not be checked
for during runtime and will not be displayed among active reminders.
Users can view what the reminder contained and if/when it was triggered,
or when it was set to (if date type reminder)
Reminder activity now has a toggle to view archived or active reminders.